Move findContainers and findSimilarElements into lib/dom.ts and some cleanup
