+++ src/lib/dom.ts
@@ -1,11 +1,17 @@
-import { equals } from 'rambdax'
+import {
+    countBy,
+    empty,
+    equals,
+    filter,
+    head,
+    identity,
+    includes,
+    sortObject,
+    tail,
+    toPairs,
+} from 'rambdax'
 
-export function sameScrollSize (nodeA, nodeB) {
-    const sameWidth = equals(nodeA.scrollWidth, nodeB.scrollWidth)
-    const sameHeight = equals(nodeA.scrollHeight, nodeB.scrollHeight)
-
-    return sameWidth && sameHeight
-}
+import { valueSorter } from '~/lib/utils'
 
 export function isVertical (elem) {
     return elem.offsetHeight > elem.offsetWidth
@@ -13,6 +19,13 @@ export function isVertical (elem) {
 
 export function isVisible (node) {
     return !!(node.offsetWidth || node.offsetHeight || node.getClientRects().length)
+}
+
+export function sameScrollSize (nodeA, nodeB) {
+    const sameWidth = equals(nodeA.scrollWidth, nodeB.scrollWidth)
+    const sameHeight = equals(nodeA.scrollHeight, nodeB.scrollHeight)
+
+    return sameWidth && sameHeight
 }
 
 // From https://youmightnotneedjquery.com/#parents
@@ -24,4 +37,46 @@ export function parents (node, selector) {
     }
 
     return parents
+}
+
+export function findContainers () {
+    const nodes = [...document.querySelectorAll(':nth-child(5)')].map(n => n.parentNode)
+
+    return nodes.filter(
+        node => isVisible(node) &&
+            isVertical(node) &&
+            !sameScrollSize(document.body, node)
+    )
+}
+
+export function findSimilarElements (container) {
+    const excludedTags = ['SCRIPT', 'IFRAME', 'STYLE']
+    const children = Array.from(container.childNodes).filter((node) => !includes(node.tagName, excludedTags))
+    const classes = children.map((node) => [node.tagName, ...node.classList])
+    const counts = countBy(identity, classes)
+    const sorted = sortObject(valueSorter, counts)
+    const top = head(toPairs(sorted))
+    const spec = (head(top) || '').split(',')
+    const tagName = head(spec)
+    const classList = tail(spec)
+
+    if (top[1] === 1) {
+        console.debug('[UFC] no similar elements found', { sorted })
+        return []
+    }
+
+    const similar = filter((node) => {
+        const sameTag = equals(tagName, node.tagName)
+        const sameClasses = equals(classList, Array.from(node.classList))
+
+        if (empty(classList)) {
+            return sameTag
+        }
+
+        return sameTag && sameClasses
+    }, children)
+
+    console.log('[UFC] findSimilarElements:', container, { similar, top, sorted, tagName, classList })
+
+    return similar
 }
